Set small number of tracking parameters through Confic class

+++ Tracking/include/TrackCandidateStrategy.hh
@@ -32,7 +32,7 @@ class DetectorGeometry;
 
 
   protected:
-    TrackCandidateStrategy(int, int);
+    TrackCandidateStrategy(int, double);
  
   public:
     TrackCandidateStrategy( const TrackCandidateStrategy&) = delete;
